Fixed all errors

+++ server/src/main/java/service/UserService.java
@@ -18,7 +18,10 @@ public class UserService {
     }
 
     public AuthData register(String username, String password, String email) throws DataAccessException {
-        if (username.isEmpty() || password.isEmpty() || email.isEmpty()) throw new DataAccessException("Error: bad request");
+        if (username == null || password == null || email == null
+            || username.isEmpty() || password.isEmpty() || email.isEmpty()) {
+            throw new DataAccessException("Error: bad request");
+        }
         if (userDAO.getUser(username) == null) {
             userDAO.createUser(new UserData(username, password, email));
             return login(username, password);
@@ -26,7 +29,10 @@ public class UserService {
     }
 
     public AuthData login(String username, String password) throws DataAccessException {
-        if (username.isEmpty() || password.isEmpty()) throw new DataAccessException("Error: bad request");
+        if (username == null || password == null
+            || username.isEmpty() || password.isEmpty()) {
+            throw new DataAccessException("Error: bad request");
+        }
         UserData user = userDAO.getUser(username);
         if (user != null && (user.username().equals(username) && user.password().equals(password))) {
             String authToken = UUID.randomUUID().toString();
@@ -38,6 +44,7 @@ public class UserService {
 
     public void logout(String authToken) throws DataAccessException {
         if (authToken.isEmpty()) throw new DataAccessException("Error: unauthorized");
+        if (authDAO.getAuth(authToken) == null) throw new DataAccessException("Error: unauthorized");
         authDAO.deleteAuth(authToken);
     }
 }
